chore(RoyalRoad): Clean-up code
